refactor(update-weight): move the transfer protocols into training_utils/weight_update
