+++ miles/backends/training_utils/weight_update/protocol.py
@@ -83,27 +83,23 @@ class WeightTransferProtocol(ABC):
 def get_weight_transfer_protocol(args: Namespace) -> WeightTransferProtocol:
     # Local: protocol modules import megatron/sglang-heavy deps.
     if args.colocate and args.update_weight_transfer_mode != "rdt":
-        from miles.backends.megatron_utils.update_weight.update_weight_from_tensor import UpdateWeightFromTensor
+        from miles.backends.training_utils.weight_update.colocate import UpdateWeightFromTensor
 
         return UpdateWeightFromTensor(args)
     if args.update_weight_transfer_mode == "broadcast":
-        from miles.backends.megatron_utils.update_weight.update_weight_from_distributed.broadcast import (
-            UpdateWeightFromDistributed,
-        )
+        from miles.backends.training_utils.weight_update.broadcast import UpdateWeightFromDistributed
 
         return UpdateWeightFromDistributed(args)
     if args.update_weight_transfer_mode == "disk-delta":
-        from miles.backends.megatron_utils.update_weight.update_weight_from_distributed.delta import (
-            UpdateWeightFromDiskDelta,
-        )
+        from miles.backends.training_utils.weight_update.delta import UpdateWeightFromDiskDelta
 
         return UpdateWeightFromDiskDelta(args)
     if args.update_weight_transfer_mode == "rdt":
-        from miles.backends.megatron_utils.update_weight.update_weight_from_rdt import UpdateWeightFromRDT
+        from miles.backends.training_utils.weight_update.rdt import UpdateWeightFromRDT
 
         return UpdateWeightFromRDT(args)
     if args.update_weight_transfer_mode == "p2p":
-        from miles.backends.megatron_utils.update_weight.update_weight_from_distributed.p2p import UpdateWeightP2P
+        from miles.backends.training_utils.weight_update.p2p import UpdateWeightP2P
 
         return UpdateWeightP2P(args)
     raise ValueError(f"Unknown --update-weight-transfer-mode {args.update_weight_transfer_mode!r}")
